fix login when secure browser is required and chromebook secure browser is used
isSecureBrowser was being used to determine the check for a required secure browser.
isSecureBrowser checks for the existence of secure browser apis, which are not present when chromebook secure browser is used.
Changing call to isSecure, which uses the browser's user agent and returns true when chromebook secure browser is used.